working word frequency ranking, added sample box
yay for a sample box, it runs slower than i would like, though, once you start getting more than a few words in then there's lag and that will need to go away in the future, i think, maybe not, if this is automated then there's not going to be a lot of typing done

I'd like to implement some basic logic rules. 
Such as if the next input contains a question mark, and the total frequency is lower than, say, .5% notify me of a possible question.