Update Main.html
Alternate ranking algorithm and organization. I don't know if this is actually better, yet.
Even if it's not faster I may prefer this due to the logic re-work.
I like the new logic a tad better, but if it comes with a huge drop in performance then it'll be dropped.